synthetic_iot_telemetry/requirements.txt
google-cloud-pubsub==2.19.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
